backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Snake Playground Pro Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "websockets>=14.0",
    "sqlalchemy>=2.0.44",
    "alembic>=1.17.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "sqlalchemy-utils>=0.42.0",
]
